Implementing new Magical Effects
Thanks to ploppy for the creativity.
