+++ EpicLoot/magiceffects.json
@@ -2859,6 +2859,48 @@
         "$mod_epicloot_me_slow_suffix1"
       ]
     },
+    {
+      "Type" : "DoubleMagicShot",
+      "DisplayText" : "$mod_epicloot_me_doublemagicshot",
+	  "Description" : "$mod_epicloot_me_doublemagicshot_desc",
+      "Requirements" : {
+        "AllowedItemTypes" : [
+          "Staff"
+		]
+      },
+      "ValuesPerRarity" : {
+        "Legendary" : { "MinValue" : 1, "MaxValue" : 1, "Increment" : 1 }
+      },
+      "SelectionWeight" : 1,
+      "Prefixes": [
+        "$mod_epicloot_me_doublemagicshot_prefix1"
+      ],
+      "Suffixes": [
+        "$mod_epicloot_me_doublemagicshot_suffix1"
+      ]
+    },
+    {
+      "Type" : "TripleBowShot",
+      "DisplayText" : "$mod_epicloot_me_triplebowshot",
+	  "Description" : "$mod_epicloot_me_triplebowshot_desc",
+      "Requirements" : {
+        "AllowedItemTypes" : [
+          "Bow"
+		]
+      },
+      "ValuesPerRarity" : {
+        "Legendary" : { "MinValue" : 1, "MaxValue" : 1, "Increment" : 1 }
+      },
+      "SelectionWeight" : 1,
+      "Prefixes": [
+        "$mod_epicloot_me_triplebowshot_prefix1"
+      ],
+      "Suffixes": [
+        "$mod_epicloot_me_triplebowshot_suffix1"
+      ]
+    }
+
+
 
     // Set Effects
     /////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -2898,26 +2940,6 @@
         "NoRoll" :  true 
       },
       "Ability" : "Berserker"
-    },
-    {
-      "Type" : "DoubleMagicShot",
-      "DisplayText" : "$mod_epicloot_set_me_doublemagicshot",
-      "Requirements" : {
-        "NoRoll" :  true 
-      },
-      "ValuesPerRarity" : {
-        "Legendary" : { "MinValue" : 1, "MaxValue" : 1, "Increment" : 1 }
-      }
-    },
-    {
-      "Type" : "TripleBowShot",
-      "DisplayText" : "$mod_epicloot_set_me_triplebowshot",
-      "Requirements" : {
-        "NoRoll" :  true 
-      },
-      "ValuesPerRarity" : {
-        "Legendary" : { "MinValue" : 1, "MaxValue" : 1, "Increment" : 1 }
-      }
     }
   ]
 }
