+++ EpicLoot/translations.json
@@ -935,7 +935,16 @@
   "mod_epicloot_me_slow_desc": "Slows the attack and movement speed of creatures being hit by +<b><color=yellow>X</color></b>% for 2 seconds. Bosses are immune to this effect.",
   "mod_epicloot_me_slow_prefix1": "Sluggish",
   "mod_epicloot_me_slow_suffix1": "Lethargy",
-
+  "mod_epicloot_me_doublemagicshot": "Double Magic Shot",
+  "mod_epicloot_me_doublemagicshot_desc": "Elemental Magic attacks fire twice, instead of once.",
+  "mod_epicloot_me_doublemagicshot_prefix1": "Elemental Doubling",
+  "mod_epicloot_me_doublemagicshot_suffix1": "of Elemental Doubling",
+  "mod_epicloot_me_triplebowshot": "Triple Bow Shot",
+  "mod_epicloot_me_triplebowshot_desc": "Bow attacks fire 3x ammo, instead of one.",
+  "mod_epicloot_me_triplebowshot_prefix1": "Tripling",
+  "mod_epicloot_me_triplebowshot_suffix1": "of Tripling",
+  
+  
   "mod_epicloot_set_me_bulwark": "Bulwark [Activated]: Prevent all damage for 5 seconds. Cooldown: 60 seconds.",
   "mod_epicloot_set_me_undying": "Undying [Passive]: On death, regain full health. Cooldown: 20 minutes.",
   "mod_epicloot_set_me_frostdamageaoe_display": "Melee attacks that deal frost damage deal an additional 50% of weapon damage as frost in a cone in front of the attacker.",
